# access_control_service/access-control-service-openapi.yaml
openapi: 3.0.3
info:
  title: Compliance & Access Control Service API
  description: |
    Микросервис для управления геозонами, правилами доступа и контроля соблюдения политик безопасности.
    Обеспечивает привязку меток к сущностям (сотрудники/оборудование), проверку разрешенных зон и генерацию оповещений о нарушениях.
  version: 1.0.0
  contact:
    name: Команда разработки
    email: security@rtls.example.com
servers:
  - url: http://localhost:8084/api/v1
    description: Локальный сервер разработки
  - url: https://compliance.rtls.example.com/api/v1
    description: Продуктивный сервер
...
paths:
...
  /entities:
    get:
      tags:
        - Entities
      summary: Получение списка всех зарегистрированных сущностей
      description: Возвращает список сотрудников и оборудования с привязанными метками.
      operationId: getAllEntities
      parameters:
        - name: entity_type
          in: query
          required: false
          schema:
            type: string
            enum: [employee, equipment, all]
            default: all
          description: Тип сущности для фильтрации
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
            maximum: 1000
          description: Максимальное количество записей
      responses:
        '200':
          description: Успешный запрос.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Entity'
        '500':
          description: Ошибка сервера.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Entities
      summary: Регистрация новой сущности
      description: Создает новую запись о сотруднике или оборудовании.
      operationId: createEntity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityCreate'
      responses:
        '201':
          description: Сущность успешно создана.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
        '400':
          description: Некорректные данные.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Конфликт (сущность с таким ID уже существует).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /entities/{entity_id}:
    get:
      tags:
        - Entities
      summary: Получение информации о конкретной сущности
      description: Возвращает детальную информацию о сущности по её ID.
      operationId: getEntityById
      parameters:
        - name: entity_id
          in: path
          required: true
          schema:
            type: string
          description: Уникальный идентификатор сущности
      responses:
        '200':
          description: Успешный запрос.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
        '404':
          description: Сущность с указанным ID не найдена.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - Entities
      summary: Обновление информации о сущности
      description: Обновляет данные сущности (включая привязку/отвязку метки).
      operationId: updateEntity
      parameters:
        - name: entity_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityUpdate'
      responses:
        '200':
          description: Сущность успешно обновлена.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
        '404':
          description: Сущность с указанным ID не найдена.
        '400':
          description: Некорректные данные.

    delete:
      tags:
        - Entities
      summary: Удаление сущности
      description: Удаляет сущность из системы.
      operationId: deleteEntity
      parameters:
        - name: entity_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Сущность успешно удалена.
        '404':
          description: Сущность с указанным ID не найдена.
  /entities/tag/{tag_id}:
    get:
      tags:
        - Entities
      summary: Получение сущности по привязанной метке
      description: Возвращает сущность, к которой привязана указанная метка.
      operationId: getEntityByTagId
      parameters:
        - name: tag_id
          in: path
          required: true
          schema:
            type: string
          description: Идентификатор метки
      responses:
        '200':
          description: Успешный запрос.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
        '404':
          description: Сущность с указанной меткой не найдена.

  # ==================== Geofences: Управление геозонами ====================
...
components:
  schemas:
...
    EntityCreate:
      type: object
      required:
        - entity_id
        - name
        - entity_type
      properties:
        entity_id:
          type: string
          description: Уникальный идентификатор сущности (например, emp-001)
        name:
          type: string
          description: Имя сущности (ФИО сотрудника или название оборудования)
        entity_type:
          type: string
          enum: [employee, equipment]
          description: Тип сущности
        tag_id:
          type: string
          description: Идентификатор метки для привязки (опционально)
        department:
          type: string
          description: Отдел/подразделение (для сотрудников)
        role:
          type: string
          description: Роль/должность (для сотрудников)
        equipment_type:
          type: string
          description: Тип оборудования (для оборудования)
        is_active:
          type: boolean
          default: true
          description: Активна ли сущность
        metadata:
          type: object
          description: Дополнительные метаданные
    EntityUpdate:
      type: object
      properties:
        name:
          type: string
        tag_id:
          type: string
          description: Привязать/отвязать метку (null для отвязки)
        department:
          type: string
        role:
          type: string
        is_active:
          type: boolean
        metadata:
          type: object
    Entity:
      type: object
      required:
        - entity_id
        - name
        - entity_type
        - created_at
      properties:
        entity_id:
          type: string
        name:
          type: string
        entity_type:
          type: string
          enum: [employee, equipment]
        tag_id:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
        equipment_type:
          type: string
          nullable: true
        is_active:
          type: boolean
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # --- Геозоны ---
...
    Error:
      type: object
      properties:
        error_code:
          type: string
          example: "ENTITY_NOT_FOUND"
        message:
          type: string
          example: "Entity with ID 'emp-001' not found"
        timestamp:
          type: string
          format: date-time
        details:
          type: object
          description: Дополнительная информация об ошибке
...